+++ otterdog/models/branch_protection_rule.py
@@ -14,10 +14,11 @@ from typing import Any, Optional
 
 from jsonbender import bend, S, OptionalS, Forall, K  # type: ignore
 
+from otterdog.jsonnet import JsonnetConfig
+from otterdog.models import ModelObject, ValidationContext, FailureType
 from otterdog.providers.github import Github
-from otterdog.utils import UNSET, is_unset, is_set_and_valid, snake_to_camel_case
-
-from . import ModelObject, ValidationContext, FailureType
+from otterdog.utils import UNSET, is_unset, is_set_and_valid, snake_to_camel_case, IndentingPrinter, \
+    write_patch_object_as_json
 
 
 @dataclasses.dataclass
@@ -257,3 +258,13 @@ class BranchProtectionRule(ModelObject):
                 mapping["requiredStatusChecks"] = K(transformed_checks)
 
         return bend(mapping, data)
+
+    def to_jsonnet(self,
+                   printer: IndentingPrinter,
+                   jsonnet_config: JsonnetConfig,
+                   extend: bool,
+                   default_object: ModelObject) -> None:
+        patch = self.get_patch_to(default_object)
+        patch.pop("pattern")
+        printer.print(f"orgs.{jsonnet_config.create_branch_protection_rule}('{self.pattern}')")
+        write_patch_object_as_json(patch, printer)